fix(users): Make password optional when editing a user
The password field in the User resource form is now only required during user creation.

When editing a user, the password field can be left blank if no change is desired. A helper text has been added to clarify this behavior. This prevents the need to re-enter the password on every edit.